parcel out validated class file processing into modules
